+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 3.3.0dev
 
+### `Added`
+
+- Added `--flag_in_source` to flag in-source fragments (`is_isf` UserParam) and re-inject them into the FDR-filtered set so their signal reaches quantification [#XXX](https://github.com/nf-core/mhcquant/pull/XXX)
+
 ### `Changed`
 
 - Updated to nf-core template 4.0.2 [#454](https://github.com/nf-core/mhcquant/pull/454)